Moving another utility to the utilities directory, reorganizing tests
